requirements.txt
annotated-types==0.5.0
anyio==3.7.1
base58==2.1.1
certifi==2023.7.22
Deprecated==1.2.14
ECPy==1.2.5
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
phonenumbers==8.13.18
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
sniffio==1.3.0
starlette==0.27.0
types-Deprecated==1.2.9.3
typing_extensions==4.7.1
websockets==10.4
wrapt==1.15.0
xrpl-py==2.2.0
charset-normalizer==3.2.0 
requests==2.31.0
urllib3==2.0.4
